Refactor annotation to a class
