+++ site/partials/examples.html
@@ -113,7 +113,7 @@ $$</p></div><div class="code"><pre><code class="javascript">functionPlot({
 <ul>
 <li><code>x</code>: x coordinate of a line parallel to the y-axis</li>
 <li><code>y</code>: y coordinate of a line parallel to the x-axis</li>
-<li><code>text</code> (optional) text shown next to the parallel line</li>
+<li><code>label</code> (optional) text shown next to the parallel line</li>
 </ul>
 <p>NOTE: either <code>x</code> or <code>y</code> need to be set on the object, setting both of them 
 will raise an exception</p></div><div class="code"><pre><code class="javascript">functionPlot({
@@ -130,11 +130,11 @@ will raise an exception</p></div><div class="code"><pre><code class="javascript"
     },
     {
       x: 1,
-      text: 'x = 1'
+      label: 'x = 1'
     },
     {
       y: 2,
-      text: 'y = 2'
+      label: 'y = 2'
     }
   ]
 })</code></pre></div></div><div class="col-md-6 center-block demos"><span class="graph" id="annotations"></span></div></div></div></div><div class="example"><div class="container"><div class="row"><div class="col-md-6"><div class="comment"><h3>Range and closed path</h3>
