+++ src/graph-types/index.ts
@@ -2,6 +2,7 @@ import { Interval, interval } from './interval.js'
 import { Polyline, polyline } from './polyline.js'
 import { Scatter, scatter } from './scatter.js'
 import { Text, text } from './text.js'
+import { Annotation, annotation } from './annotation.js'
 import { Attr, Mark } from './mark.js'
 
-export { Polyline, polyline, Scatter, scatter, Interval, interval, Text, text, Attr, Mark }
+export { Polyline, polyline, Scatter, scatter, Interval, interval, Text, text, Annotation, annotation, Attr, Mark }
